requirements.txt
pandas
python-telegram-bot
requests
hyper
openpyxl